requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
firebase-admin==6.5.0
razorpay==1.4.1
httpx==0.27.0
redis==5.0.4
Pillow==10.3.0
python-slugify==8.0.4
aiofiles==23.2.1
psycopg2-binary==2.9.9
greenlet==3.0.3
